Adding command line functions, duplications, and file dumps

+++ include/simulations/urnlearning_simulation.h
@@ -6,6 +6,6 @@
 
 typedef double ** (*payoff_function)(unsigned int players, unsigned int **types, unsigned int * state_action_profile);
 
-void urnlearning_dynamics(urngame_t *urngame, unsigned int max_iterations, payoff_function payoffs);
+void urnlearning_dynamics(urngame_t *urngame, unsigned long max_iterations, payoff_function payoffs);
 
 #endif
